+++ src/iaduban_movements.c
@@ -14,8 +14,8 @@
 
 void		takeoff_player(t_context *context)
 {
-	int x;
-	int y;
+	size_t	x;
+	size_t	y;
 
 	x = 0;
 	y = 0;
@@ -34,8 +34,8 @@ void		takeoff_player(t_context *context)
 
 void		place_player(t_context *context)
 {
-	int x;
-	int y;
+	size_t	x;
+	size_t	y;
 
 	x = 0;
 	y = 0;
@@ -54,11 +54,11 @@ void		place_player(t_context *context)
 
 void		move_x(t_context *context, int x_target)
 {
-	if (x_target > context->player.pos.x &&
+	if (x_target > (int)context->player.pos.x &&
 			context->map[context->player.pos.y][context->player.pos.x + 1]
 			== MAP_EMPTYCASE)
 		context->player.pos.x += 1;
-	else if (x_target < context->player.pos.x &&
+	else if (x_target < (int)context->player.pos.x &&
 			context->map[context->player.pos.y][context->player.pos.x - 1]
 			== MAP_EMPTYCASE)
 		context->player.pos.x -= 1;
@@ -69,11 +69,11 @@ void		move_x(t_context *context, int x_target)
 
 void		move_y(t_context *context, int y_target)
 {
-	if (y_target > context->player.pos.y &&
+	if (y_target > (int)context->player.pos.y &&
 			context->map[context->player.pos.y + 1][context->player.pos.x]
 			== MAP_EMPTYCASE)
 		context->player.pos.y += 1;
-	else if (y_target < context->player.pos.y &&
+	else if (y_target < (int)context->player.pos.y &&
 			context->map[context->player.pos.y - 1][context->player.pos.x]
 			== MAP_EMPTYCASE)
 		context->player.pos.y -= 1;
